+++ src/main/java/com/example/demo/repository/RentRepository.java
@@ -3,12 +3,13 @@ package com.example.demo.repository;
 import com.example.demo.entity.Rent;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface RentRepository extends CrudRepository<Rent,Integer> {
+public interface RentRepository extends CrudRepository<Rent,Long> {
+
     List<Rent> findAll();
-    Optional<Rent> findRentById(Integer rentId);
+    Optional<Rent> findRentById(Long rentId);
+
 }
