+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -3,11 +3,13 @@ package com.example.demo.repository;
 import com.example.demo.entity.User;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
+
 @Repository
-public interface UserRepository extends CrudRepository<User,Integer> {
+public interface UserRepository extends CrudRepository<User,Long> {
+
     List<User> findAll();
-    Optional<User> findUserById(Integer userId);
+    Optional<User> findUserById(Long userId);
+
 }
